complete

+++ src/main/java/vn/tlcn/trungtamgiasu/service/ClassesService.java
@@ -17,6 +17,7 @@ import vn.tlcn.trungtamgiasu.dto.SearchDto;
 import vn.tlcn.trungtamgiasu.dto.mapper.ClassesMapper;
 import vn.tlcn.trungtamgiasu.exception.ClassesNotFoundException;
 import vn.tlcn.trungtamgiasu.exception.NotChangeStatusClass;
+import vn.tlcn.trungtamgiasu.model.ClassRegister;
 import vn.tlcn.trungtamgiasu.model.Classes;
 import vn.tlcn.trungtamgiasu.model.Tutors;
 import vn.tlcn.trungtamgiasu.model.Users;
@@ -188,9 +189,20 @@ public class ClassesService {
         return classesRepository.findAll();
     }
 
-    public void deleteClass(int id){
-        classesRepository.deleteById(id);
+    public boolean deleteClass(int id){
+        boolean flag = true;
+        List<ClassRegister> classRegisters = classRegisterRepository.findAll();
+        for (ClassRegister classRegister: classRegisters) {
+            if(classRegister.getClasses().getIdClass() == id){
+                flag = false;
+            }
+        }
+        if(flag == true){
+            classesRepository.deleteById(id);
+        }
+        return flag;
     }
+
 
     public Users getParent(Classes classes){
         return classes.getUsers();
